+++ src/ServiceFabric/ServiceFabric.Test/ScenarioTests/Common.ps1
@@ -192,6 +192,37 @@ function WaitForClusterReadyState($clusterName, $resourceGroupName, $timeoutInSe
     return $false
 }
 
+function WaitForManagedClusterReadyStateIfRecord($clusterName, $resourceGroupName)
+{
+	if ([Microsoft.Azure.Test.HttpRecorder.HttpMockServer]::Mode -ne [Microsoft.Azure.Test.HttpRecorder.HttpRecorderMode]::Playback)
+	{
+		# Wait for Ready cluster state before updating otherwise update is going to fail
+		if (-not (WaitForManagedClusterReadyState $clusterName $resourceGroupName))
+		{
+			Assert-True $false 'Cluster is not in Ready state. Can not continue with test.'
+		}
+	}
+}
+
+function WaitForManagedClusterReadyState($clusterName, $resourceGroupName, $timeoutInSeconds = 1200)
+{
+    $timeoutTime = (Get-Date).AddSeconds($timeoutInSeconds)
+    while (-not $clusterReady -and (Get-Date) -lt $timeoutTime) {
+        $cluster = (Get-AzServiceFabricManagedCluster -ResourceGroupName $resourceGroupName -Name $clusterName)[0]
+        if ($cluster.ClusterState -eq "Ready")
+        {
+            return $true
+            break
+        }
+
+        Write-Host "Cluster state: $($cluster.ClusterState). Waiting for Ready state before continuing."
+        Start-Sleep -Seconds 15
+    }
+
+    Write-Error "WaitForClusterReadyState timed out"
+    return $false
+}
+
 function WaitForAllJob($timeoutInSeconds = 1200)
 {
     $timeoutTime = (Get-Date).AddSeconds($timeoutInSeconds)
@@ -217,6 +248,53 @@ function WaitForAllJob($timeoutInSeconds = 1200)
 
     Write-Error "WaitForJob timed out"
     return $false
+}
+
+<#
+.SYNOPSIS
+Asserts if two hashtables with simple key and value types are equal
+#>
+function Assert-HashtableEqual($h1, $h2)
+{
+  if($h1.count -ne $h2.count)
+  {
+    throw "Hashtable size not equal. Hashtable1: " + $h1.count + " Hashtable2: " + $h2.count
+  }
+
+  foreach($key in $h1.Keys)
+  {
+    if($h1[$key] -ne $h2[$key])
+    {
+      throw "Tag content not equal. Key:$key Tags1:" +  $h1[$key] + " Tags2:" + $h2[$key]
+    }
+  }
+}
+
+###################
+#
+# Verify that the actual string ends with the expected suffix
+#
+#    param [string] $expectedSuffix : The expected suffix
+#    param [string] $actual         : The actual string
+#    param [string] $message        : The message to return if the actual string does not end with the suffix
+####################
+function Assert-EndsWith
+{
+    param([string] $expectedSuffix, [string] $actual, [string] $message)
+
+  Assert-NotNull $actual
+
+  if (!$message)
+  {
+      $message = "Assertion failed because actual '$actual' does not end with '$expectedSuffix'"
+  }
+
+  if (-not $actual.EndsWith($expectedSuffix))
+  {
+      throw $message
+  }
+
+  return $true
 }
 
 # Application functions
@@ -249,4 +327,41 @@ function Get-AppPackageV2
 function Get-ServiceTypeName
 {
     return "CalcServiceType"
+}
+
+# Managed Application functions
+
+function Get-ManagedAppTypeName
+{
+    return "VotingType"
+}
+
+function Get-ManagedAppTypeV1Name
+{
+    return "1.0.0"
+}
+
+function Get-ManagedAppTypeV2Name
+{
+    return "2.0.0"
+}
+
+function Get-ManagedAppPackageV1
+{
+    return "https://sfmconeboxst.blob.core.windows.net/managed-application-deployment/Voting.sfpkg"
+}
+
+function Get-ManagedAppPackageV2
+{
+    return "https://sfmconeboxst.blob.core.windows.net/managed-application-deployment/Voting.2.0.0.sfpkg"
+}
+
+function Get-ManagedStatelessServiceTypeName
+{
+    return "VotingWebType"
+}
+
+function Get-ManagedStatefulServiceTypeName
+{
+    return "VotingDataType"
 }