+++ src/Accounts/Accounts/ChangeLog.md
@@ -19,6 +19,10 @@
 -->
 ## Upcoming Release
 
+## Version 2.2.7
+* Fixed incorrect warning message on Windows PowerShell [#14556]
+* Set Azure Environment variable `AzureKeyVaultServiceEndpointResourceId` according to the value of `AzureKeyVaultDnsSuffix` when discovering environment 
+
 ## Version 2.2.6
 * Upgrade Azure.Identity to fix the issue that Connect-AzAccount fails when ADFS credential is used [#13560]